master: multi_checkpoint.cpp: remove an unecessary conversion int->size_t.

+++ test_more/cppad_for_tmb/multi_checkpoint.cpp
@@ -86,7 +86,7 @@ bool multi_checkpoint(void)
 		v         = f.Forward(0, x);
 		//
 		// this assigment has false sharing; i.e., will case cache resets
-		y[ size_t(thread) ] = v[0];
+		y[thread] = v[0];
 	}
 
 	// check the results
